requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
python-multipart==0.0.9
PyPDF2==3.0.1
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
pytest==8.3.4
